refactor: update import paths for SectionTitle component
