+++ src/common/components/SectionTitle.tsx
