+++ src/pages/Campaign/useWidgets/Experience/widgets/General/index.tsx
@@ -1,10 +1,10 @@
 import { Campaign } from 'src/features/api';
 import { Col, Grid, Row } from '@appquality/unguess-design-system';
-import { SectionTitle } from 'src/pages/Campaign/SectionTitle';
 import { Divider } from 'src/common/components/divider';
 import { appTheme } from 'src/app/theme';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
+import { SectionTitle } from 'src/common/components/SectionTitle';
 import { GoalCard } from './GoalCard';
 import { Methodology } from './Methodology';
